requirements.txt
streamlit>=1.44.0
openai>=1.75.0
python-dotenv>=1.0.1
supabase>=2.5.1
stripe>=12.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
